Initial commit
The Asana API key is now expected to be an environment variable, so the
code itself doesn't have any private information hard-coded into it. For the
script to be able to connect to the Asana servers it's expecting to
read an environment variable named 'ASANA_KEY' on whatever computer
it's currently running on. This private key should only be available to
computers which have that variable set (i.e. only the lab computers).
It cannot be gathered from this source code alone and without the key
this script is non-functional.